requirements.txt
#pandas==2.1.2
pandas
ipykernel==6.26.0
tqdm
rdkit
fastparquet
mhfp==1.9.6
neomodel==5.2.1
seaborn==0.13.1
imbalanced-learn==0.11.0
#scikit-learn==1.3.2
scikit-learn
venny4py==1.0.1
pycaret==3.2.0
ipywidgets==8.1.1
torch==2.1.2
optuna==3.5.0
openpyxl==3.1.2
matplotlib_venn==0.11.10
treeinterpreter
xgboost